refactor : 사다리 테스트 파일 패키지 분리
